+++ sdk/servicebus/azure-servicebus/azure/servicebus/_transport/_pyamqp_transport.py
@@ -658,7 +658,11 @@ class PyamqpTransport(AmqpTransport):   # pylint: disable=too-many-public-method
         """
         # pylint: disable=protected-access
         message = message_type(
-            message=received[1], receive_mode=receiver._receive_mode, receiver=receiver, frame=received[0]
+            message=received[1],
+            receive_mode=receiver._receive_mode,
+            receiver=receiver,
+            frame=received[0],
+            amqp_transport=receiver._amqp_transport
         )
         receiver._last_received_sequenced_number = message.sequence_number
         return message
@@ -852,7 +856,9 @@ class PyamqpTransport(AmqpTransport):   # pylint: disable=too-many-public-method
         condition: Optional["ErrorCondition"],
         description: str,
         exception: Optional["AMQPException"] = None,
-        status_code: Optional[str] = None
+        status_code: Optional[str] = None,
+        *,
+        custom_endpoint_address: Optional[str] = None
     ) -> "ServiceBusError":
         # handling AMQP Errors that have the condition field or the mgmt handler
         logger.info(
@@ -864,7 +870,11 @@ class PyamqpTransport(AmqpTransport):   # pylint: disable=too-many-public-method
         error_cls: Type["ServiceBusError"]
         if isinstance(exception, AuthenticationException):
             logger.info("AMQP Connection authentication error occurred: (%r).", exception)
-            error_cls = ServiceBusAuthenticationError
+            if custom_endpoint_address:
+                # for uamqp parity, invalid custom endpoint address raises ServiceBusConnectionError
+                error_cls = ServiceBusConnectionError
+            else:
+                error_cls = ServiceBusAuthenticationError
         # elif isinstance(exception, MessageException):
         #     logger.info("AMQP Message error occurred: (%r).", exception)
         #     if isinstance(exception, MessageAlreadySettled):
@@ -920,14 +930,18 @@ class PyamqpTransport(AmqpTransport):   # pylint: disable=too-many-public-method
 
     @staticmethod
     def create_servicebus_exception(
-        logger: "Logger", exception: Exception
+        logger: "Logger", exception: Exception, *, custom_endpoint_address: Optional[str] = None
     ) -> "ServiceBusError":
         if isinstance(exception, AMQPException):
             # handling AMQP Errors that have the condition field
             condition = exception.condition
             description = exception.description
             exception = PyamqpTransport._handle_amqp_exception_with_condition(
-                logger, condition, description, exception=exception
+                logger,
+                condition,
+                description,
+                exception=exception,
+                custom_endpoint_address=custom_endpoint_address
             )
         elif not isinstance(exception, ServiceBusError):
             logger.exception(