+++ sdk/servicebus/azure-servicebus/tests/async_tests/test_sb_client_async.py
@@ -4,12 +4,17 @@
 # license information.
 #--------------------------------------------------------------------------
 
-
 import logging
+import sys
 import time
+import asyncio
 import pytest
+import hmac
+import hashlib
+import base64
+from urllib.parse import quote as url_parse_quote
 
-from azure.core.credentials import AzureSasCredential, AzureNamedKeyCredential
+from azure.core.credentials import AzureSasCredential, AzureNamedKeyCredential, AccessToken
 from azure.mgmt.servicebus.models import AccessRights
 from azure.servicebus.aio import ServiceBusClient, ServiceBusSender, ServiceBusReceiver
 from azure.servicebus import ServiceBusMessage
@@ -17,7 +22,8 @@ from azure.servicebus.aio._base_handler_async import ServiceBusSharedKeyCredenti
 from azure.servicebus.exceptions import (
     ServiceBusError,
     ServiceBusAuthenticationError,
-    ServiceBusAuthorizationError
+    ServiceBusAuthorizationError,
+    ServiceBusConnectionError
 )
 from devtools_testutils import AzureMgmtRecordedTestCase
 from servicebus_preparer import (
@@ -343,6 +349,45 @@ class TestServiceBusClientAsync(AzureMgmtRecordedTestCase):
             async with client.get_queue_sender(servicebus_queue.name) as sender:
                 await sender.send_messages(ServiceBusMessage("foo"))
 
+        def generate_sas_token(uri, sas_name, sas_value, token_ttl):
+            """Performs the signing and encoding needed to generate a sas token from a sas key."""
+            sas = sas_value.encode('utf-8')
+            expiry = str(int(time.time() + token_ttl))
+            string_to_sign = (uri + '\n' + expiry).encode('utf-8')
+            signed_hmac_sha256 = hmac.HMAC(sas, string_to_sign, hashlib.sha256)
+            signature = url_parse_quote(base64.b64encode(signed_hmac_sha256.digest()))
+            return 'SharedAccessSignature sr={}&sig={}&se={}&skn={}'.format(uri, signature, expiry, sas_name)
+
+        class CustomizedSASCredential(object):
+            def __init__(self, token, expiry):
+                """
+                :param str token: The token string
+                :param float expiry: The epoch timestamp
+                """
+                self.token = token
+                self.expiry = expiry
+                self.token_type = b"servicebus.windows.net:sastoken"
+
+            async def get_token(self, *scopes, **kwargs):
+                """
+                This method is automatically called when token is about to expire.
+                """
+                return AccessToken(self.token, self.expiry)
+
+        token_ttl = 5  # seconds
+        sas_token = generate_sas_token(
+            auth_uri, servicebus_namespace_key_name, servicebus_namespace_primary_key, token_ttl
+        )
+        credential=CustomizedSASCredential(sas_token, time.time() + token_ttl)
+
+        async with ServiceBusClient(hostname, credential, uamqp_transport=uamqp_transport) as client:
+            sender = client.get_queue_sender(queue_name=servicebus_queue.name)
+            await asyncio.sleep(10)
+            with pytest.raises(ServiceBusAuthenticationError):
+                async with sender:
+                    message = ServiceBusMessage("Single Message")
+                    await sender.send_messages(message)
+
     @pytest.mark.asyncio
     @pytest.mark.liveTest
     @pytest.mark.live_test_only
@@ -591,15 +636,15 @@ class TestServiceBusClientAsync(AzureMgmtRecordedTestCase):
             assert subscription_receiver.client_identifier is not None
             assert subscription_receiver.client_identifier == custom_id
 
-    @pytest.mark.skip('check that connection verify works for pyproto. Issue #26657.')
     @pytest.mark.asyncio
     @pytest.mark.liveTest
+    @pytest.mark.live_test_only
     @CachedServiceBusResourceGroupPreparer()
     @CachedServiceBusNamespacePreparer(name_prefix='servicebustest')
     @CachedServiceBusQueuePreparer(name_prefix='servicebustest')
     @pytest.mark.parametrize("uamqp_transport", uamqp_transport_params, ids=uamqp_transport_ids)
     @ArgPasserAsync()
-    async def test_connection_verify_exception_async(self,
+    async def test_custom_endpoint_connection_verify_exception_async(self,
                                    uamqp_transport,
                                    *,
                                    servicebus_queue=None,
@@ -617,3 +662,30 @@ class TestServiceBusClientAsync(AzureMgmtRecordedTestCase):
                 async with client.get_queue_sender(servicebus_queue.name) as sender:
                     await sender.send_messages(ServiceBusMessage("foo"))
 
+        # Skipping on OSX uamqp - it's raising an Authentication/TimeoutError
+        if not uamqp_transport or not sys.platform.startswith('darwin'):
+            fake_addr = "fakeaddress.com:1111"
+            client = ServiceBusClient(
+                hostname,
+                credential,
+                custom_endpoint_address=fake_addr,
+                retry_total=0,
+                uamqp_transport=uamqp_transport
+            )
+            async with client:
+                with pytest.raises(ServiceBusConnectionError):
+                    async with client.get_queue_sender(servicebus_queue.name) as sender:
+                        await sender.send_messages(ServiceBusMessage("foo"))
+
+            client = ServiceBusClient(
+                hostname,
+                credential,
+                custom_endpoint_address=fake_addr,
+                connection_verify="cacert.pem",
+                retry_total=0,
+                uamqp_transport=uamqp_transport,
+            )
+            async with client:
+                with pytest.raises(ServiceBusError):
+                    async with client.get_queue_sender(servicebus_queue.name) as sender:
+                        await sender.send_messages(ServiceBusMessage("foo"))