+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -62,7 +62,7 @@
 
 Version 7.9.0a1 is our first efforts to build an Azure Service Bus client library based on a pure Python implemented AMQP stack.
 
-### Breaking changes
+### Breaking Changes
 
 - The following features have been temporarily pulled out which will be added back in future previews as we work towards a stable release:
   - Iterator receiving from Service Bus entities.